+++ test/helpers/functionsMetaData.spec.ts
@@ -1,8 +1,8 @@
 import { readJSON } from 'fs-extra'
 import mock from 'mock-fs'
 import { join } from 'pathe'
-import { NEXT_PLUGIN_NAME } from '../packages/runtime/src/constants'
-import { writeFunctionConfiguration } from '../packages/runtime/src/helpers/functionsMetaData'
+import { NEXT_PLUGIN_NAME } from '../../packages/runtime/src/constants'
+import { writeFunctionConfiguration } from '../../packages/runtime/src/helpers/functionsMetaData'
 
 describe('writeFunctionConfiguration', () => {
   afterEach(() => {
