+++ test/templates/server.spec.ts
@@ -1,8 +1,8 @@
 import { mockRequest } from 'next/dist/server/lib/mock-request'
 import { Options } from 'next/dist/server/next-server'
 
-import { getNextServer, NextServerType, netlifyApiFetch } from '../packages/runtime/src/templates/handlerUtils'
-import { NetlifyNextServer, NetlifyConfig } from '../packages/runtime/src/templates/server'
+import { getNextServer, NextServerType, netlifyApiFetch } from '../../packages/runtime/src/templates/handlerUtils'
+import { NetlifyNextServer, NetlifyConfig } from '../../packages/runtime/src/templates/server'
 
 jest.mock('../packages/runtime/src/templates/handlerUtils', () => {
   const originalModule = jest.requireActual('../packages/runtime/src/templates/handlerUtils')
